+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Makefile: Add coverage rule to check a local HTML report
 - agents: Closed-loop model predictive control
 - envs: Base `Reward` class for rewards
+- envs: Survival reward, which is simply always one
 - examples: Closed-loop model predictive control
 - utils: Configure agent process on the Raspberry Pi
 
@@ -17,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 - Makefile: Remove agent targets to promote running their ``main.py``
 - README: Recommended way to run agents is now via Python
+- envs: Default reward for all environments is now the survival reward
 - examples: Remove CPU isolation example, now a ``utils.raspi`` function call
 - utils: Remove ``realtime`` submodule in favor of ``raspi``